# hot-wheels-backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Hot Wheels Collector API
  version: 1.0.0
  description: |
    # Hot Wheels Collector Platform API

    A comprehensive REST API for Hot Wheels collectors to manage their collections,
    trade with other collectors, and explore the complete Hot Wheels catalog.

    ## Features
    - **Authentication**: JWT-based secure authentication system
    - **Catalog**: Browse 50+ years of Hot Wheels models
    - **Collections**: Track and manage personal collections
    - **Marketplace**: Buy, sell, and trade with other collectors
    - **Reviews**: Rate and review transactions

    ## Getting Started
    1. Register a new account via `/api/v1/auth/register`
    2. Login to receive JWT tokens via `/api/v1/auth/login`
    3. Include the access token in the `Authorization` header
    4. Start exploring the catalog and managing your collection!

    ## Rate Limits
    - Anonymous users: 100 requests/minute
    - Authenticated users: 500 requests/minute
    - Rate limit headers included in all responses

    ## Support
    For API support, please contact the development team.

  contact:
    name: API Support
    email: support@hotwheels-collector.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.hotwheels-collector.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from `/api/v1/auth/login`.
        Include in Authorization header: `Bearer <token>`
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: "Invalid request parameters"
            code:
              type: string
              example: "VALIDATION_ERROR"
...
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: "hotwheels_fan"
        full_name:
          type: string
          nullable: true
          example: "John Smith"
        country:
          type: string
          nullable: true
          example: "United States"
        city:
          type: string
          nullable: true
          example: "Los Angeles"
        profile_image_url:
          type: string
          nullable: true
          example: "https://example.com/avatar.jpg"
        bio:
          type: string
          nullable: true
          example: "Collecting Hot Wheels since 1985"
        is_verified:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        role:
          type: string
          enum: [collector, trader, admin, moderator]
          example: "collector"
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        last_login:
          type: string
          format: date-time
          nullable: true
          example: "2024-12-17T08:00:00Z"
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: "hotwheels_fan"
        email:
          type: string
          format: email
          example: "collector@example.com"
        password:
          type: string
          minLength: 8
          maxLength: 128
          format: password
          example: "SecureP@ss123"
        full_name:
          type: string
          nullable: true
          example: "John Smith"
        country:
          type: string
          nullable: true
          example: "United States"
        city:
          type: string
          nullable: true
          example: "Los Angeles"
        bio:
          type: string
          nullable: true
          example: "Passionate Hot Wheels collector"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "collector@example.com"
        password:
          type: string
          format: password
          example: "SecureP@ss123"
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 86400
          description: Token expiration time in seconds
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
          nullable: true
          example: "John Smith Jr."
        country:
          type: string
          nullable: true
          example: "Canada"
        city:
          type: string
          nullable: true
          example: "Toronto"
        bio:
          type: string
          nullable: true
          example: "Updated bio"
        profile_image_url:
          type: string
          nullable: true
          example: "https://example.com/new-avatar.jpg"
    ChangePasswordRequest:
      type: object
      required:
        - old_password
        - new_password
      properties:
        old_password:
          type: string
          format: password
          example: "OldP@ss123"
        new_password:
          type: string
          minLength: 8
          format: password
          example: "NewSecureP@ss456"

    # ==================== Catalog Schemas ====================
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Missing required fields: email, password"
    Unauthorized:
      description: Unauthorized - authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Authentication required"
...
    Conflict:
      description: Conflict - resource already exists or invalid state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              message: "Email already registered"
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user
      description: |
        Creates a new user account with the provided credentials.

        **Password Requirements:**
        - Minimum 8 characters
        - Maximum 128 characters

        **Username Requirements:**
        - 3-50 characters
        - Must be unique
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              basic:
                summary: Basic registration
                value:
                  username: "hotwheels_fan"
                  email: "collector@example.com"
                  password: "SecureP@ss123"
              full:
                summary: Full registration with all fields
                value:
                  username: "hotwheels_fan"
                  email: "collector@example.com"
                  password: "SecureP@ss123"
                  full_name: "John Smith"
                  country: "United States"
                  city: "Los Angeles"
                  bio: "Collecting Hot Wheels since 1985"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: "User registered successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      summary: User login
      description: |
        Authenticates a user and returns JWT access and refresh tokens.

        **Token Usage:**
        - Access token expires in 24 hours
        - Refresh token expires in 30 days
        - Include access token in `Authorization: Bearer <token>` header
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "collector@example.com"
              password: "SecureP@ss123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AuthTokens'
                  message:
                    type: string
                    example: "Login successful"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      summary: User logout
      description: Logs out the current user and invalidates the session
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      summary: Update user profile
      description: Updates the profile of the currently authenticated user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                  message:
                    type: string
                    example: "Profile updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/change-password:
    post:
      summary: Change password
      description: Changes the password for the currently authenticated user
      tags:
        - Authentication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Password changed successfully. Please login again."
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      description: |
        Returns a new access token using the refresh token.

        Include the refresh token in the Authorization header.
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  message:
                    type: string
                    example: "Token refreshed successfully"
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== Catalog ====================
...